fixes fixes fixes. also cache is 2gb so I didn't push it
