--- nids_clustering/corpus/readme.txt
@@ -1,14 +1,0 @@
-The dataset is known as the KDDÕ99 NIDS dataset. 
-
-The dataset contains network traffic samples (as flows) generated from active network attacks and baseline activity. 
-
-Further details and the original, unprocessed data can be found here:
-
-http://kdd.ics.uci.edu/databases/kddcup99/kddcup99.html
-
-
-The KDDÕ99 dataset that you have been given has been preprocessed into normalized features and divided into training and testing splits. 
-
-The training split contains only baseline traffic samples and is provided in the file named training_normal.npy.
-
-The testing data is organized into two files: one for testing on unseen baseline traffic (named testing_normal.npy), and another for testing on anomalous malicious traffic (named testing_attack.npy).
