--- nids_clustering/docs/requirements.txt
@@ -1,3 +1,0 @@
-numpy
-pandas
-scikit-learn
